Fix recursive driver.Valuer implementations

+++ service/persist/contract.go
@@ -48,7 +48,7 @@ const (
 
 // Value implements the driver.Valuer interface for the Chain type
 func (c ContractOwnerMethod) Value() (driver.Value, error) {
-	return c, nil
+	return int64(c), nil
 }
 
 // Scan implements the sql.Scanner interface for the Chain type
